simple move of MainWindow to its own file
