+++ gui/electron/windows/GraphWindow.js
